backend/pyproject.toml
[project]
name = "milo-backend"
version = "0.1.0"
description = "Milo backend - FastAPI services"
readme = "README.md"
requires-python = ">=3.11"
authors = [ { name = "Warren LATA", email = "latawarren@gmail.com" } ]



dependencies = [
  "fastapi>=0.98.0",
  "uvicorn[standard]>=0.22.0",
  "pydantic>=2.5.1",
  "python-multipart>=0.0.5",
  "requests>=2.32.0",
  "google-adk>=1.28.1",
  "vertexai>=1.43.0",
  "pre-commit>=4.5.1",
  "firebase-admin>=7.3.0",
  "sql>=2022.4.0",
  "sqlalchemy>=2.0.49",
  "asyncpg>=0.31.0",
  "cloud-sql-python-connector>=1.20.1",
  "pyrebase4>=4.9.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"



[tool.ruff]
line-length = 88
extend-exclude = [".venv", "migrations", "__pycache__"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88
